+++ src/deephaven_mcp/sessions/_session_manager.py
